rename build to avoid name conflict
